+++ Framework/AssemblyInfo.cs
@@ -1,4 +1,4 @@
 using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("Unity.InternalAPIEditorBridgeDev.005")]
-[assembly: InternalsVisibleTo("Assembly-CSharp-Editor-firstpass-testable")]
+[assembly: InternalsVisibleTo("Unity.InternalAPIEditorBridgeDev.002")]