+++ Framework/Interactive Tutorials/Editor/Internal Proxy/AssemblyInfo.cs.cs
@@ -1,4 +1,3 @@
 using System.Runtime.CompilerServices;
 
-[assembly: InternalsVisibleTo("Assembly-CSharp-Editor-firstpass-testable")]
-[assembly: InternalsVisibleTo("Unity.learn.iet-framework.Editor.Tests")]
+[assembly: InternalsVisibleTo("Unity.InternalAPIEditorBridgeDev.002")]